POM and prototype POM fixed Things moved from POM to prototype POM.
POM fitted with just project data now.
Prototype POM has dependency, compiler, surefire config and the like.

TODO: testng.xml, sources to fix, install locally and test if it works.

It has been said that the great scientific disciplines are examples of giants standing on the shoulders of other giants.  It has also been said that the software industry is an example of midgets standing on the toes of other midgets.
      -- Alan Cooper